Revert "IQuantity with QuantityValue as Value type (PoC)"
This reverts commit eae2fd55f6366514b016172a3417e43b8f071123.

+++ UnitsNet.Serialization.JsonNet/AbbreviatedUnitsConverter.cs
@@ -85,7 +85,7 @@ namespace UnitsNet.Serialization.JsonNet
             }
             else
             {
-                writer.WriteValue((double)quantity.Value);
+                writer.WriteValue(quantity.Value);
             }
 
             //  write the 'Unit' abbreviation
